# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StatusHub API",
    "version": "1.0.0",
    "description": "API REST para monitorización de servicios IT en tiempo real.",
    "contact": {
      "email": "admin@statushub.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor local (Docker)"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Introduce el token obtenido en /api/auth/login"
      }
    }
  },
  "paths": {
...
    "/api/incidents": {
      "get": {
        "tags": ["Incidentes"],
        "summary": "Lista de incidentes",
        "description": "Devuelve la lista paginada de incidentes. Permite filtrar por servicio y estado.",
        "parameters": [
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "description": "Filtrar por ID de servicio",
            "schema": { "type": "integer", "example": 1 }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filtrar por estado",
            "schema": {
              "type": "string",
              "enum": ["investigating", "identified", "monitoring", "resolved"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de incidentes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "current_page": { "type": "integer", "example": 1 },
                    "total": { "type": "integer", "example": 10 },
                    "per_page": { "type": "integer", "example": 10 },
                    "data": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Incidentes"],
        "summary": "Crear un incidente",
        "description": "Crea un nuevo incidente y actualiza el estado del servicio según el impacto. Solo administradores.",
        "security": [{ "sanctum": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["service_id", "title", "impact", "status"],
                "properties": {
                  "service_id": { "type": "integer", "example": 1 },
                  "title": { "type": "string", "example": "Latencia elevada en API de Usuarios" },
                  "description": { "type": "string", "example": "Se detecta un aumento inusual en los tiempos de respuesta." },
                  "impact": { "type": "string", "enum": ["minor", "major", "critical"], "example": "minor" },
                  "status": { "type": "string", "enum": ["investigating", "identified", "monitoring", "resolved"], "example": "investigating" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Incidente creado correctamente" },
          "401": { "description": "No autenticado" },
          "403": { "description": "No autorizado" }
        }
      }
    },
    "/api/incidents/{incident}": {
      "get": {
        "tags": ["Incidentes"],
        "summary": "Detalle de un incidente",
        "description": "Devuelve los detalles de un incidente incluyendo su servicio y actualizaciones.",
        "parameters": [
          {
            "name": "incident",
            "in": "path",
            "required": true,
            "description": "ID del incidente",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": { "description": "Detalle del incidente obtenido correctamente" },
          "404": { "description": "Incidente no encontrado" }
        }
      },
      "put": {
        "tags": ["Incidentes"],
        "summary": "Actualizar un incidente",
        "description": "Actualiza un incidente. Si se resuelve, restaura el servicio si no quedan incidentes activos. Solo administradores.",
        "security": [{ "sanctum": [] }],
        "parameters": [
          {
            "name": "incident",
            "in": "path",
            "required": true,
            "description": "ID del incidente",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "title": { "type": "string", "example": "Latencia elevada en API de Usuarios" },
                  "description": { "type": "string", "example": "Se detecta un aumento inusual en los tiempos de respuesta." },
                  "impact": { "type": "string", "enum": ["minor", "major", "critical"] },
                  "status": { "type": "string", "enum": ["investigating", "identified", "monitoring", "resolved"], "example": "resolved" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Incidente actualizado correctamente" },
          "401": { "description": "No autenticado" },
          "403": { "description": "No autorizado" }
        }
      },
      "delete": {
        "tags": ["Incidentes"],
        "summary": "Eliminar un incidente",
        "description": "Elimina un incidente y restaura el servicio si no quedan incidentes activos. Solo administradores.",
        "security": [{ "sanctum": [] }],
        "parameters": [
          {
            "name": "incident",
            "in": "path",
            "required": true,
            "description": "ID del incidente",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Incidente eliminado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Incidente eliminado correctamente." }
                  }
                }
              }
            }
          },
          "401": { "description": "No autenticado" },
          "403": { "description": "No autorizado" }
        }
      }
    },
    "/api/incidents/{incident}/updates": {
      "get": {
        "tags": ["Actualizaciones"],
        "summary": "Lista de actualizaciones de un incidente",
        "description": "Devuelve todas las actualizaciones de un incidente específico.",
        "parameters": [
          {
            "name": "incident",
            "in": "path",
            "required": true,
            "description": "ID del incidente",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": { "description": "Lista de actualizaciones obtenida correctamente" },
          "404": { "description": "Incidente no encontrado" }
        }
      },
      "post": {
        "tags": ["Actualizaciones"],
        "summary": "Añadir actualización a un incidente",
        "description": "Crea una nueva actualización y sincroniza el estado del incidente. Solo administradores.",
        "security": [{ "sanctum": [] }],
        "parameters": [
          {
            "name": "incident",
            "in": "path",
            "required": true,
            "description": "ID del incidente",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["message", "status"],
                "properties": {
                  "message": { "type": "string", "example": "Hemos identificado la causa raíz." },
                  "status": { "type": "string", "enum": ["investigating", "identified", "monitoring", "resolved"], "example": "identified" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Actualización creada correctamente" },
          "401": { "description": "No autenticado" },
          "403": { "description": "No autorizado" }
        }
      }
    },
...
  }
}